+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-## Next
+## 1.1.0 (5/1/2014)
 
 * Added `ARTiledImageScrollView.backgroundImage` - [@dblock](https://github.com/dblock).
 * Exposed `ARTile`, drawing now goes through `ARTile#drawInRect` - [@dblock](https://github.com/dblock).
 
-## 1.0.0 (3/15/2013)
+## 1.0.0 (3/15/2014)
 
 * Initial public release - [@dblock](https://github.com/dblock).